Replace Handlebars config variable with helper
Extension to #1401. In that PR I originally introduced the `config`
helper, but it was preferred that SFRPG not mix the `config` variable
and a `config` helper. It was suggested that a different PR could
introduce the helper and replace all the variables at once.

I took the opportunity to also assimilate and extend the behavior of
`lookup` into the helper, as `(lookup config.<value> variable)` was a
common pattern.